SAS Routes: Generalised params func and added sorting
